+++ oshi-core/src/test/java/oshi/SystemInfoTest.java
@@ -119,11 +119,6 @@ public class SystemInfoTest {
         assertNotNull(p.getSystemSerialNumber());
         assertTrue(p.getLogicalProcessorCount() >= p.getPhysicalProcessorCount());
         assertTrue(p.getPhysicalProcessorCount() > 0);
-        assertTrue(p.getProcessCount() >= 1);
-        assertTrue(p.getThreadCount() >= 1);
-        assertTrue(p.getProcessId() > 0);
-        assertNotNull(p.getProcesses());
-        assertEquals(p.getProcess(p.getProcessId()).getProcessID(), p.getProcessId());
     }
 
     /**
@@ -236,7 +231,7 @@ public class SystemInfoTest {
     }
 
     /**
-     * Test get version.
+     * Test operating system
      */
     @Test
     public void testOSVersion() {
@@ -249,6 +244,12 @@ public class SystemInfoTest {
         assertNotNull(version.getVersion());
         assertNotNull(version.getCodeName());
         assertNotNull(version.getBuildNumber());
+
+        assertTrue(os.getProcessCount() >= 1);
+        assertTrue(os.getThreadCount() >= 1);
+        assertTrue(os.getProcessId() > 0);
+        assertNotNull(os.getProcesses());
+        assertEquals(os.getProcess(os.getProcessId()).getProcessID(), os.getProcessId());
     }
 
     /**
@@ -260,10 +261,10 @@ public class SystemInfoTest {
     @Test
     public void testFileSystem() throws IOException {
         SystemInfo si = new SystemInfo();
-        FileSystem filesystem = si.getHardware().getFileSystem();
+        FileSystem filesystem = si.getOperatingSystem().getFileSystem();
         assertTrue(filesystem.getOpenFileDescriptors() >= 0L);
         assertTrue(filesystem.getMaxFileDescriptors() >= 0L);
-        OSFileStore[] fs = si.getHardware().getFileStores();
+        OSFileStore[] fs = filesystem.getFileStores();
         for (int f = 0; f < fs.length; f++) {
             assertNotNull(fs[f].getName());
             assertNotNull(fs[f].getVolume());
@@ -293,12 +294,12 @@ public class SystemInfoTest {
 
         LOG.info("Initializing System...");
         SystemInfo si = new SystemInfo();
+
         // software
         // software: operating system
         OperatingSystem os = si.getOperatingSystem();
         System.out.println(os);
 
-        LOG.info("Initializing Hardware...");
         // hardware
         HardwareAbstractionLayer hal = si.getHardware();
 
@@ -356,9 +357,8 @@ public class SystemInfoTest {
         System.out.println(procCpu.toString());
 
         // Processes
-        System.out.println("Processes: " + hal.getProcessor().getProcessCount() + ", Threads: "
-                + hal.getProcessor().getThreadCount());
-        List<OSProcess> procs = Arrays.asList(hal.getProcessor().getProcesses());
+        System.out.println("Processes: " + os.getProcessCount() + ", Threads: " + os.getThreadCount());
+        List<OSProcess> procs = Arrays.asList(os.getProcesses());
         // Sort by highest CPU
         Comparator<OSProcess> cpuDescOrder = new Comparator<OSProcess>() {
             @Override
@@ -384,7 +384,6 @@ public class SystemInfoTest {
                     FormatUtil.formatBytes(p.getVirtualSize()), FormatUtil.formatBytes(p.getResidentSetSize()),
                     p.getName());
         }
-
         // hardware: sensors
         LOG.info("Checking Sensors...");
         System.out.println("Sensors:");
@@ -413,24 +412,6 @@ public class SystemInfoTest {
         }
         System.out.println(sb.toString());
 
-        // hardware: file system
-        LOG.info("Checking File System...");
-        System.out.println("File System:");
-
-        FileSystem filesystem = hal.getFileSystem();
-        System.out.format(" File Descriptors: %d/%d%n", filesystem.getOpenFileDescriptors(),
-                filesystem.getMaxFileDescriptors());
-
-        OSFileStore[] fsArray = hal.getFileStores();
-        for (OSFileStore fs : fsArray) {
-            long usable = fs.getUsableSpace();
-            long total = fs.getTotalSpace();
-            System.out.format(" %s (%s) [%s] %s of %s free (%.1f%%) is %s and is mounted at %s%n", fs.getName(),
-                    fs.getDescription().isEmpty() ? "file system" : fs.getDescription(), fs.getType(),
-                    FormatUtil.formatBytes(usable), FormatUtil.formatBytes(fs.getTotalSpace()), 100d * usable / total,
-                    fs.getVolume(), fs.getMount());
-        }
-
         // hardware: disks
         LOG.info("Checking Disks...");
         System.out.println("Disks:");
@@ -443,6 +424,24 @@ public class SystemInfoTest {
                     dsk.getSize() > 0 ? FormatUtil.formatBytesDecimal(dsk.getSize()) : "?",
                     readwrite ? FormatUtil.formatBytes(dsk.getReads()) : "?",
                     readwrite ? FormatUtil.formatBytes(dsk.getWrites()) : "?");
+        }
+
+        // software: file system
+        LOG.info("Checking File System...");
+        System.out.println("File System:");
+
+        FileSystem filesystem = os.getFileSystem();
+        System.out.format(" File Descriptors: %d/%d%n", filesystem.getOpenFileDescriptors(),
+                filesystem.getMaxFileDescriptors());
+
+        OSFileStore[] fsArray = filesystem.getFileStores();
+        for (OSFileStore fs : fsArray) {
+            long usable = fs.getUsableSpace();
+            long total = fs.getTotalSpace();
+            System.out.format(" %s (%s) [%s] %s of %s free (%.1f%%) is %s and is mounted at %s%n", fs.getName(),
+                    fs.getDescription().isEmpty() ? "file system" : fs.getDescription(), fs.getType(),
+                    FormatUtil.formatBytes(usable), FormatUtil.formatBytes(fs.getTotalSpace()), 100d * usable / total,
+                    fs.getVolume(), fs.getMount());
         }
 
         // hardware: network interfaces