+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 3.0 (in progress)
 ================
-* [#205](https://github.com/dblock/oshi/pull/190): Separated oshi-core and oshi-json modules - [@dbwiddis](https://github.com/dbwiddis).
+* [#205](https://github.com/dblock/oshi/pull/205): Separated oshi-core and oshi-json modules - [@dbwiddis](https://github.com/dbwiddis).
+* [#209](https://github.com/dblock/oshi/pull/209): Move FileSystem and Processes to OperatingSystem - [@dbwiddis](https://github.com/dbwiddis).
 * Your contribution here.
+
 
 2.6.2 (6/21/2016)
 ================